html for admin dashboard
the basic html for the admin dashboard page is created.
the page is not styled yet. but is divided up using the grid.
its two columns, the left contains the sidebar with menu items,
and the right contains the main content of the page. its beginning to rain.
i am going to commit this now.